pyproject.toml
[project]
name = "drug-response-prediction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "biopython>=1.85",
    "gseapy>=1.1.9",
    "jupyter>=1.1.1",
    "lifelines>=0.30.0",
    "matplotlib>=3.10.6",
    "numpy>=2.3.2",
    "pandas>=2.3.2",
    "plotly>=6.3.0",
    "scikit-learn>=1.7.1",
    "seaborn>=0.13.2",
    "statsmodels>=0.14.5",
    "streamlit>=1.49.1",
]
